refactor contacts
