+++ handlers/src/github.rs
@@ -1,0 +1,6 @@
+gh_pages_handlers!(
+    [
+        contacts_handler,
+        "contact-info.json",
+        types::contacts::Response
+    ],
