+++ routes/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod contacts;
